refat:  mudança no esquema do banco

+++ backend/Caritas.Models/Entities/Endereco.cs
@@ -7,17 +7,13 @@ namespace Caritas.Models.Entities;
 public class Endereco
 {
     [Key]
-    public int Id { get; set; }
-
+    public long Id { get; set; }
     public string? Rua { get; set; }
-
     public string? Numero { get; set; }
-
     public string? Cep { get; set; }
-
     public string? Bairro { get; set; }
-
     public string? Cidade { get; set; }
 
+    // Navegação inversa
     public Paroquia? Paroquia { get; set; }
 }
